LCD16x2Display.js - Started
So i started working on the lcd display today. It will be hard, but its not so imposible. I think cases will be nut, but it is definitly possible to make it finish tomorrow. On the other way i find problem with inheritance on the objects. Maybe fix later. but i dont thinks so